requirements.txt
easyocr>=1.7.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
